restructure data mod
