+++ src/authorization.rs
@@ -8,8 +8,8 @@ pub use vercre_did::{Resource, dereference};
 use vercre_infosec::jose::JwsBuilder;
 use vercre_infosec::{Jws, Signer};
 
-use crate::data::cid;
 use crate::records::DelegatedGrant;
+use crate::utils::cid;
 use crate::{Result, unexpected};
 
 /// Generate a closure to resolve pub key material required by `Jws::decode`.
