+++ src/clients/protocols.rs
@@ -6,7 +6,7 @@
 use chrono::{DateTime, Utc};
 
 use crate::authorization::AuthorizationBuilder;
-use crate::data::cid;
+use crate::utils::cid;
 pub use crate::protocols::{
     self, Configure, ConfigureDescriptor, Definition, ProtocolsFilter, Query, QueryDescriptor,
 };
@@ -66,10 +66,10 @@ impl ConfigureBuilder {
         let mut definition = self.definition.ok_or_else(|| unexpected!("definition not found"))?;
 
         // normalize definition urls
-        definition.protocol = utils::clean_url(&definition.protocol)?;
+        definition.protocol = utils::uri::clean(&definition.protocol)?;
         for t in definition.types.values_mut() {
             if let Some(schema) = &t.schema {
-                t.schema = Some(utils::clean_url(schema)?);
+                t.schema = Some(utils::uri::clean(schema)?);
             }
         }
         protocols::validate_structure(&definition)?;
