+++ src/permissions/grant.rs
@@ -457,7 +457,7 @@ impl Grant {
 //         // N.B. adding a protocol tag ensures message queries with a protocol
 //         // filter will return associated grants
 //         if let Some(protocol) = scope.protocol() {
-//             let protocol = utils::clean_url(protocol)?;
+//             let protocol = utils::uri::clean(protocol)?;
 //             builder = builder.add_tag("protocol".to_string(), Value::String(protocol));
 //         };
 
@@ -541,7 +541,7 @@ impl Grant {
 //         // N.B. adding a protocol tag ensures message queries with a protocol
 //         // filter will return this request
 //         if let Some(protocol) = scope.protocol() {
-//             let protocol = utils::clean_url(protocol)?;
+//             let protocol = utils::uri::clean(protocol)?;
 //             builder = builder.add_tag("protocol".to_string(), Value::String(protocol));
 //         };
 
@@ -605,7 +605,7 @@ impl Grant {
 //         // N.B. adding a protocol tag ensures message queries with a protocol
 //         // filter will return this request
 //         if let Some(protocol) = grant_data.scope.protocol() {
-//             let protocol = utils::clean_url(protocol)?;
+//             let protocol = utils::uri::clean(protocol)?;
 //             builder = builder.add_tag("protocol".to_string(), Value::String(protocol));
 //         };
