+++ src/protocols/query.rs
@@ -3,7 +3,7 @@
 use serde::{Deserialize, Serialize};
 
 use crate::authorization::Authorization;
-use crate::data::cid;
+use crate::utils::cid;
 use crate::endpoint::{Message, Reply, Status};
 use crate::protocols::{Configure, ProtocolsFilter};
 use crate::provider::{MessageStore, Provider};
@@ -26,7 +26,7 @@ pub async fn handle(
 ) -> Result<Reply<QueryReply>> {
     // validate query
     if let Some(filter) = &query.descriptor.filter {
-        utils::validate_url(&filter.protocol)?;
+        utils::uri::validate(&filter.protocol)?;
     }
 
     // build actual query
