+++ src/records/write.rs
@@ -15,15 +15,17 @@ use vercre_infosec::Signer;
 use vercre_infosec::jose::{Jws, JwsBuilder};
 
 use crate::authorization::{Authorization, JwsPayload};
-use crate::data::cid;
 use crate::endpoint::{Message, Reply, Status};
 use crate::permissions::{self, Grant, Protocol};
 use crate::protocols::{PROTOCOL_URI, REVOCATION_PATH, integrity};
 use crate::provider::{DataStore, EventLog, EventStream, MessageStore, Provider};
 use crate::records::{DateRange, EncryptionProperty};
 use crate::serde::{rfc3339_micros, rfc3339_micros_opt};
-use crate::store::{Entry, EntryType, GrantedQueryBuilder, RecordsFilter, RecordsQueryBuilder};
-use crate::{Descriptor, Error, Method, Result, authorization, data, forbidden, unexpected};
+use crate::store::{
+    Entry, EntryType, GrantedQueryBuilder, RecordsFilter, RecordsQueryBuilder, data,
+};
+use crate::utils::cid;
+use crate::{Descriptor, Error, Method, Result, authorization, forbidden, unexpected};
 
 /// Handle `RecordsWrite` messages.
 ///
