+++ src/store/data.rs
@@ -1,0 +1,8 @@
+//! Data record handling.
+
+/// The maximum size of a message.
+pub const MAX_ENCODED_SIZE: usize = 30000;
+
+/// The maximum size of a block.
+pub const CHUNK_SIZE: usize = 16;
+
