+++ tests/records_read.rs
@@ -6,13 +6,14 @@ use base64ct::{Base64UrlUnpadded, Encoding};
 use dwn_node::clients::grants::GrantBuilder;
 use dwn_node::clients::protocols::{ConfigureBuilder, QueryBuilder};
 use dwn_node::clients::records::{Data, DeleteBuilder, ProtocolBuilder, ReadBuilder, WriteBuilder};
-use dwn_node::data::{MAX_ENCODED_SIZE, cid};
+use dwn_node::store::MAX_ENCODED_SIZE;
 use dwn_node::hd_key::{self, DerivationPath, DerivationScheme, DerivedPrivateJwk, PrivateKeyJwk};
 use dwn_node::permissions::{RecordsScope, Scope};
 use dwn_node::protocols::Definition;
 use dwn_node::provider::{BlockStore, MessageStore};
 use dwn_node::records::{EncryptOptions, Recipient, RecordsFilter, decrypt};
 use dwn_node::store::Entry;
+use dwn_node::cid;
 use dwn_node::{Error, Method, endpoint};
 use http::StatusCode;
 use rand::RngCore;
