+++ tests/records_write.rs
@@ -10,14 +10,14 @@ use dwn_node::clients::protocols::ConfigureBuilder;
 use dwn_node::clients::records::{
     Data, DeleteBuilder, ProtocolBuilder, QueryBuilder, ReadBuilder, WriteBuilder,
 };
-use dwn_node::data::MAX_ENCODED_SIZE;
+use dwn_node::store::MAX_ENCODED_SIZE;
 use dwn_node::hd_key::{DerivationScheme, PrivateKeyJwk};
 use dwn_node::messages::MessagesFilter;
 use dwn_node::permissions::{Conditions, Publication, RecordsScope, Scope};
 use dwn_node::protocols::{Definition, ProtocolType, RuleSet, Size};
 use dwn_node::provider::EventLog;
 use dwn_node::records::{Attestation, EncryptOptions, Recipient, RecordsFilter, SignaturePayload};
-use dwn_node::{Error, Interface, Message, Method, clients, data, endpoint, store};
+use dwn_node::{Error, Interface, Message, Method, clients, endpoint, store};
 use http::StatusCode;
 use rand::RngCore;
 use test_node::key_store::{
@@ -5665,7 +5665,7 @@ async fn context_id_mismatch() {
     // alter the signature context ID
     let payload = SignaturePayload {
         base: JwsPayload {
-            descriptor_cid: data::cid::from_value(&write.descriptor).unwrap(),
+            descriptor_cid: dwn_node::cid::from_value(&write.descriptor).unwrap(),
             ..JwsPayload::default()
         },
         record_id: write.record_id.clone(),
@@ -5705,7 +5705,7 @@ async fn invalid_attestation() {
     // alter the signature attestation CID
     let payload = SignaturePayload {
         base: JwsPayload {
-            descriptor_cid: data::cid::from_value(&write.descriptor).unwrap(),
+            descriptor_cid: dwn_node::cid::from_value(&write.descriptor).unwrap(),
             ..JwsPayload::default()
         },
         record_id: write.record_id.clone(),
@@ -5753,7 +5753,7 @@ async fn attestation_descriptor_cid() {
 
     // alter the attestation descriptor_cid
     let payload = Attestation {
-        descriptor_cid: data::cid::from_value(&"somerandomrecordid").expect("should create CID"),
+        descriptor_cid: dwn_node::cid::from_value(&"somerandomrecordid").expect("should create CID"),
     };
     let attestation = JwsBuilder::new()
         .payload(payload)
@@ -5764,12 +5764,12 @@ async fn attestation_descriptor_cid() {
 
     let payload = SignaturePayload {
         base: JwsPayload {
-            descriptor_cid: data::cid::from_value(&write.descriptor).unwrap(),
+            descriptor_cid: dwn_node::cid::from_value(&write.descriptor).unwrap(),
             ..JwsPayload::default()
         },
         record_id: write.record_id.clone(),
         context_id: write.context_id.clone(),
-        attestation_cid: Some(data::cid::from_value(&attestation).unwrap()),
+        attestation_cid: Some(dwn_node::cid::from_value(&attestation).unwrap()),
         ..SignaturePayload::default()
     };
     write.authorization.signature = JwsBuilder::new()
